requirements.txt
fastapi
fastapi[all]
uvicorn
python-dotenv
passlib[bcrypt]
pydantic
aiosmtplib
email-validator
google-cloud-firestore
python-jose
pyjwt
python-multipart
pydantic-settings
python-jose[cryptography]
requests
firebase-admin
google-generativeai>=0.3.0
google-api-core>=2.0.0